heavy: add error checks, fix strict aliasing and linux
The core problem was the cuda hefty Thread per block set to high
but took me several hours to find that...

btw... +25% in heavy 12500 with 256 threads per block... vs 128 & 512
if max reg count is set to 80...

+++ stats.cpp
@@ -37,7 +37,7 @@ void stats_remember_speed(int thr_id, uint32_t hashcount, double hashrate, uint8
 		return;
 
 	memset(&data, 0, sizeof(data));
-	data.uid = uid;
+	data.uid = (uint32_t) uid;
 	data.gpu_id = gpu;
 	data.thr_id = (uint8_t)thr_id;
 	data.tm_stat = (uint32_t) time(NULL);
